Add cluster info to cluster data
