+++ openshift_cli_installer/libs/osd_clusters.py
@@ -1,6 +1,12 @@
 import click
 from ocm_python_wrapper.cluster import Cluster, Clusters
 
+from openshift_cli_installer.utils.helpers import (
+    add_cluster_info_to_cluster_data,
+    dump_cluster_data_to_file,
+    get_cluster_object,
+)
+
 
 def osd_check_existing_clusters(clusters, ocm_client):
     for _cluster in Clusters(client=ocm_client).get():
@@ -35,6 +41,12 @@ def osd_create_cluster(cluster_data):
             channel_group=cluster_data["channel-group"],
             expiration_time=cluster_data.get("expiration-time"),
         )
+        cluster_data = add_cluster_info_to_cluster_data(
+            cluster_data=cluster_data,
+            cluster_object=get_cluster_object(cluster_data=cluster_data),
+        )
+        dump_cluster_data_to_file(cluster_data=cluster_data)
+
     except Exception as ex:
         click.secho(
             f"Failed to run cluster create for cluster {cluster_data['name']}\n{ex}",
